requirements.txt
discord.py[voice]
youtube_dl
black
ffmpeg
PyNaCl
python-decouple